For subtask BSAP-637 "Wettkampftabelle in 4 Tabellen für jeden Wettkampftag aufteilen"
~ wettkampf.component.html: Changed the id of the tables to table1, table2, table3, table4
these tables are only shown if new variable in wettkampf.component.ts this.showAll is true.
+ wettkampf.component.ts: Added new method getTitle() in wettkampf.component.ts to change the title of table1 if "alle anzeigen" is selected.

For Subtask BSAPP-638
"Button für "alle anzeigen" sollte nur einmal betätigt werden müssen"

wettkampf.component.ts: Added new variable showAll, in wettkampf.component.html the variable loadAll is set true if the first button is clicked. If you click on "alle anzeigen" after that showAll is set to true so that the button works correctly on the first click.
In onSelect() showAll is set to false to make sure the tables disappear if you change the Veranstaltung