New Pop-Up Message for deleting the Auffuellmannschaft Verein.
A Pop-Up message has been created that is displayed when you try to delete the Verein that is used to create the Auffuellmannschaft. At the moment you can still delete this club, but a different Pop-up message will be displayed than for the other deletion attempts for the other Vereine.